+++ src/com/missionhub/broadcast/PersonJsonSqlBroadcast.java
@@ -10,28 +10,28 @@ public class PersonJsonSqlBroadcast extends MissionHubBroadcast {
 	public static final String NOTIFY_PERSON_DELETE = MissionHubBroadcast.PREFIX + "NOTIFY_PERSON_DELETE";
 	public static final String NOTIFY_PERSON_ERROR  = MissionHubBroadcast.PREFIX + "NOTIFY_PERSON_ERROR";
 
-	public static void broadcastCreate(Context context, int personId, String ... categories) {
+	public static void broadcastCreate(Context context, long personId, String ... categories) {
 		final Intent intent = new Intent(NOTIFY_PERSON_CREATE);
 		intent.putExtra(MissionHubBroadcast.PREFIX + "personId", personId);
 		addCategories(intent, categories);
 		sendBroadcast(context, intent);
 	}
 	
-	public static void broadcastUpdate(Context context, int personId, String ... categories) {
+	public static void broadcastUpdate(Context context, long personId, String ... categories) {
 		final Intent intent = new Intent(NOTIFY_PERSON_UPDATE);
 		intent.putExtra(MissionHubBroadcast.PREFIX + "personId", personId);
 		addCategories(intent, categories);
 		sendBroadcast(context, intent);
 	}
 	
-	public static void broadcastDelete(Context context, int personId, String ... categories) {
+	public static void broadcastDelete(Context context, long personId, String ... categories) {
 		final Intent intent = new Intent(NOTIFY_PERSON_DELETE);
 		intent.putExtra(MissionHubBroadcast.PREFIX + "personId", personId);
 		addCategories(intent, categories);
 		sendBroadcast(context, intent);
 	}
 	
-	public static void broadcastError(Context context, int personId, Throwable t, String ... categories) {
+	public static void broadcastError(Context context, long personId, Throwable t, String ... categories) {
 		final Intent intent = new Intent(NOTIFY_PERSON_ERROR);
 		intent.putExtra(MissionHubBroadcast.PREFIX + "personId", personId);
 		intent.putExtra(MissionHubBroadcast.PREFIX + "throwable", t);
